pyproject.toml
[project]
name = "project_advanced_db"
version = "0.1.0"
description = ""
authors = [
    {name = "kietxuan",email = "kietnguyenxuan2401@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "neo4j (>=6.1.0,<7.0.0)",
    "pydantic (>=2.13.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "faker (>=40.13.0,<41.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
